Improving the stability of overriding existent attributes
Methods/classmethods/staticmethods created in the C-Land have different free()
functions from the same versions created from the Python-Land.

So, this commit is an attempt to skip the hassle of messing up with the `base`
free and inject custom code. Here we are just moving the old object to a `_c_`
prefixed attribute.