Fix usage of dir() without arg to inspect local scope:
A call to dir() without arguments is supposed to list the names in the
local scope. Because forbiddenfruit monkey patches __buitin__.dir with
its own __filtered_dir__, an additional frame is present in the call
stack, making CPythons _dir_locals_ access the wrong frame.

This change makes __filtered_dir__ account for that and return
the names from the local scope of the actual calling frame.